requirements.txt
# Core data processing and analysis
pandas>=1.0.0

# Database connections
clickhouse_connect>=0.5.0

# Environment variables
python-dotenv>=0.19.0

# HTTP requests
requests>=2.25.0

# Data visualization
matplotlib>=3.3.0

# Network analysis
networkx>=2.5.0

# Jupyter notebooks
jupyter>=1.0.0
notebook>=6.0.0
ipykernel>=6.0.0

# Hugging Face Hub
huggingface_hub>=0.0.12
